Mover componentes de Home a componente HomePage y agregar Router
