+++ DREAM Lab Frontend/package.json
@@ -15,15 +15,14 @@
     "@fortawesome/free-regular-svg-icons": "^6.5.1",
     "@fortawesome/free-solid-svg-icons": "^6.5.1",
     "@fortawesome/react-fontawesome": "^0.2.0",
-    "autoprefixer": "^10.4.18",
     "bootstrap": "^5.3.3",
     "clsx": "^2.1.0",
     "embla-carousel-react": "^8.0.0",
     "framer-motion": "^11.0.8",
-    "postcss": "^8.4.35",
     "react": "^18.2.0",
     "react-bootstrap": "^2.10.1",
     "react-dom": "^18.2.0",
+    "react-router-dom": "^6.22.3",
     "tailwind-merge": "^2.2.1"
   },
   "devDependencies": {
