+++ DREAM Lab Frontend/src/Home/ImageSlider.css
@@ -30,6 +30,7 @@
 	width: 100%;
 	object-fit: cover;
 	transform: scale(1.4);
+	user-select: none;
 }
 
 .embla__parallax {
